toggleCodeView
